DAEDALUS model for health, social and economic costs of a pandemic

Run the DAEDALUS model from R. This is a work in progress.

Usage

daedalus(country, ..., time_end = 300)

Arguments

country

A country or territory name from among country_names.

...

Additional arguments passed to make_parameters() or make_initial_state(). See Details for permitted arguments.

time_end

An integer-like value for the number of timesteps at which to return data. This is treated as the number of days with data returned for each day. Defaults to 300 days.

Value

A <deSolve> object.

Details

DAEDALUS allows users to substitute default model parameters with custom values by passing them as ... (see examples). These arguments are passed on to the internal functions make_parameters() and make_initial_state() as appropriate. Any arguments that are passed and are not suitable for those functions will throw a warning.

Initial state

make_initial_state() accepts the following parameters:

p_infectious: A single numeric value in the range \((0.0, 1.0)\) giving the proportion of individuals in each age group and economic sector that are to be initialised as infectious. Defaults to 1e-6, or one in every one million as infectious.

p_asymptomatic: A single numeric value in the range \((0.0, 1.0)\) for the proportion of initially infectious individuals who are considered to be asymptomatic. Defaults to 0.0.

Model parameters

make_parameters() allows users to pass the following pathogen parameters:

beta: A single numeric value > 0.0 for the infection transmission rate \(\beta\). Defaults to 0.0186, assuming an \(R_0\) of 1.3 and an infectious period (\(T_I\)) of 7.0 days (for \(\beta = R_0 / T_I\)).

sigma: A single numeric value > 0.0 for the rate of transition from the exposed compartment to one of two infectious compartments. Defaults to 0.333 assuming a pre-infectious period of 3.0 days.

p_sigma: A single numeric value in the range \((0.0, 1.0)\) for the proportion of infectious individuals who are also symptomatic. Asymptomatic individuals can have a different contribution to the force of infection from symptomatic individuals. Defaults to 0.66.

epsilon: A single numeric value for the relative contribution of asymptomatic infectious individuals to the force of infection (compared to symptomatic individuals). Defaults to 0.2.

gamma: A single numeric value for the recovery rate. Defaults to 0.143, assuming an infectious period of 7.0 days.

eta: A single numeric value for the hospitalisation rate of symptomatic infectious individuals. Defaults to 0.01, assuming one of every hundred symptomatic individuals needs hospitalisation.

omega: A single numeric value for the mortality rate of hospitalised individuals. Defaults to 0.01, assuming one in every hundred hospitalisations results in death.

rho: A single numeric value for the rate at which infection-derived immunity wanes, returning individuals in the 'recovered' compartment to the 'susceptible' compartment. Defaults to 0.00556, assuming that such immunity lasts on average 180 days.

demography: A numeric vector of length N_AGE_GROUPS (4), giving the number of individuals in each age group of interest. By default, this is taken from country_data based on the country selected.

contact_matrix: A square numeric matrix with N_AGE_GROUPS rows and columns, giving the per-capita number of contacts between individuals of each age group. By default, this is taken from country_data based on country.

contacts_workplace: A numeric vector of length N_ECON_SECTORS (45), giving the per-capita number of worker-to-worker interactions within economic sectors. Defaults to a standard value provided by the package as economic_contacts.

contacts_consumer_worker: A numeric matrix with N_ECON_SECTORS rows (45) and N_AGE_GROUPS columns (4) giving the per-capita contacts between individuals in each economic sector and consumers. Defaults to the sector-wise workplace contacts distributed in proportion to country demography.

contacts_between_sectors: A square numeric matrix with N_ECON_SECTORS rows and columns and with a zero diagonal, giving the number of worker-to-worker contacts across economic sectors. Defaults to the package standard of a null matrix as between-sector-contacts are not currently modelled.