daedalus: Model Health, Social, and Economic Costs of a Pandemic

Model the health, education, and economic costs of directly transmitted respiratory virus pandemics, under different scenarios of prior vaccine investment and reactive interventions, using the 'DAEDALUS' integrated health-economics model adapted from Haw et al. (2022) <doi.org/10.1038/s43588-022-00233-0>.

See also

Author

Maintainer: Pratik Gupte p.gupte24@imperial.ac.uk (ORCID)

Authors:

Patrick Doohan (ORCID)

Robert Johnson (ORCID)

Pablo Perez-Guzman (ORCID)

Rich FitzJohn (ORCID)

Emma Russell

David Mears

Katharina Hauck (ORCID)

Other contributors:

Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]

Imperial College of Science, Technology and Medicine [copyright holder, funder]