DAEDALUS model implemented with dust

daedalus::daedalus2() will eventually replace daedalus::daedalus() with the model implemented in C++ to integrate with dust2 (and in future dust). This is a work in progress!

Usage

daedalus2(time_end = 100, ...)

Arguments

time_end

An integer-like value for the number of timesteps at which to return data. This is treated as the number of days with data returned for each day. Defaults to 300 days.

...

Optional named parameters for the model. See Details for more.

Details

Optional parameters

Pass any of the following as named values (see example). This is a work in progress.

"I0": Initial number of exposed.

"N": Initial population size assumed to be uniform for all strata.

"beta": The transmission rate of the infection.

"sigma": The infectiousness rate.

"gamma": The recovery rate.

"n_strata": The number of strata (e.g. demography groups).

Examples

daedalus2(5, n_strata = 4)
#> $S
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> [1,]  990 989.8201 989.3413 988.6334 987.7442 986.7062
#> [2,]  990 989.8201 989.3413 988.6334 987.7442 986.7062
#> [3,]  990 989.8201 989.3413 988.6334 987.7442 986.7062
#> [4,]  990 989.8201 989.3413 988.6334 987.7442 986.7062
#> 
#> $E
#>      [,1]   [,2]     [,3]     [,4]     [,5]     [,6]
#> [1,]   10 8.3555 7.278765 6.604056 6.215337 6.031465
#> [2,]   10 8.3555 7.278765 6.604056 6.215337 6.031465
#> [3,]   10 8.3555 7.278765 6.604056 6.215337 6.031465
#> [4,]   10 8.3555 7.278765 6.604056 6.215337 6.031465
#> 
#> $I
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> [1,]    0 1.733519 3.047128 4.071878 4.899892 5.596013
#> [2,]    0 1.733519 3.047128 4.071878 4.899892 5.596013
#> [3,]    0 1.733519 3.047128 4.071878 4.899892 5.596013
#> [4,]    0 1.733519 3.047128 4.071878 4.899892 5.596013
#> 
#> $R
#>      [,1]       [,2]      [,3]     [,4]     [,5]     [,6]
#> [1,]    0 0.09085237 0.3327677 0.690692 1.140615 1.666295
#> [2,]    0 0.09085237 0.3327677 0.690692 1.140615 1.666295
#> [3,]    0 0.09085237 0.3327677 0.690692 1.140615 1.666295
#> [4,]    0 0.09085237 0.3327677 0.690692 1.140615 1.666295
#> 
#> $cases_inc
#>      [,1]      [,2]      [,3]      [,4]      [,5]     [,6]
#> [1,]    0 0.1798713 0.4787894 0.7079653 0.8892183 1.037928
#> [2,]    0 0.1798713 0.4787894 0.7079653 0.8892183 1.037928
#> [3,]    0 0.1798713 0.4787894 0.7079653 0.8892183 1.037928
#> [4,]    0 0.1798713 0.4787894 0.7079653 0.8892183 1.037928
#>