DAEDALUS system of ODEs

A compartmental epidemiological model with four age groups. This function is intended to be passed to the solvers from the deSolve package. This is a work in progress and additional components will be added in future.

Usage

daedalus_rhs(t, state, parameters)

Arguments

t

A single double value for the time.

state

Vector of the initial state. The form of the vector should be \(X_i, for X \in S, E, I_s, I_a, R, H, D\), where values for each of the compartments for each age group \(i\) are given consecutively. May be a matrix to allow for multiple age groups. See daedalus() for an explanation of the epidemiological compartments.

parameters

List for the parameters used in the simulation. See daedalus() for a list of model parameters.

Value

A list with a single numeric vector of the same size as state, suitable as output for deSolve functions such as deSolve::lsoda().

Details: DAEDALUS model

This section is a work in progress, and describes the DAEDALUS model in brief.

The model is age-stratified, supports heterogeneity in social contacts between age groups, and currently supports four age groups.

Epidemiological model

The model has the following epidemiological compartments: susceptible (\(S\)), exposed (\(E\)), infectious and symptomatic (\(I_s\)), infectious asymptomatic (\(I_a\)), hospitalised (\(H\)), recovered (\(R\)), and dead (\(D\)).

The model currently allows only uniform parameters for transitions between compartments. Group specific parameters are expected to be supported in future.