Model epidemic and economic outcomes under bespoke interventions

Allows modelling of epidemic trajectories and economic costs similar to daedalus(), but with the start and duration of restrictions controlled by the user. No reactive interventions are allowed, and vaccination is not implemented. Offers the option to model spontaneous social distancing, and allows exogeneous reduction of the transmission rate to simulate the effect of measures that reduce transmission (such as social distancing or mask mandates).

Also allows basic modelling of parameter uncertainty by passing a list of infection parameter combinations, see infection below.

Usage

daedalus_rtm(
  country,
  infection,
  response_strategy = NULL,
  response_time_start = 0,
  response_time_end = 0,
  hospital_capacity = NULL,
  initial_state_manual = list(p_infectious = 1e-07),
  auto_social_distancing = FALSE,
  social_distancing_mandate = 1,
  time_end = 300
)

Arguments

country

A country or territory object of class <daedalus_country>, or a country or territory name from those included in the package; see country_names, or a country ISO2 or ISO3 code; see country_codes_iso2c and country_codes_iso3c . Country-specific data such as the community and workplace contacts, the demography, and the distribution of the workforce into economic sectors is automatically accessed from package data for the relevant country name if it is passed as a string. To override package defaults for country characteristics, pass a <daedalus_country> object instead. See daedalus_country() for more.

infection

Similar to daedalus(), may be a character vector from among epidemic_names or a <daedalus_infection> object. May also be a list of <daedalus_infection> objects.

response_strategy

Either a string for the name of response strategy followed from among "none", "school_closures", "economic_closures", and "elimination", or a numeric vector of the same length as the number of economic sectors (45) giving the openness coefficient for each sector when closures are active.

response_time_start

A single number for the time at which the response_strategy comes into effect.

response_time_end

A single number of the time at which the response comes to an end.

hospital_capacity

A single number specifying the hospital capacity dedicated to pandemic response. When hospital demand crosses this value, the infection parameter omega is increased by a factor of 1.6.

initial_state_manual

An optional named list with the names p_infectious and p_asymptomatic for the proportion of infectious and symptomatic individuals in each age group and economic sector. Defaults to 1e-7 and 0.0 respectively.

auto_social_distancing

A single logical value for whether spontaneous social distancing is active. When active, social distancing is dependent on daily deaths and is active even when no response is active. Defaults to FALSE.

social_distancing_mandate

A single number for the scaling of the transmission rate due to other measures taken to reduce transmission, such as mask mandates. This scaling is only active between response_time_start and response_time_end, and is active even when no response strategy is specified. Defaults to 1.0.

time_end

The end point of the simulation, defaults to 300 days.