Get epidemic costs from a DAEDALUS model run

Get epidemic costs from a DAEDALUS model run

Usage

get_costs(
  x,
  summarise_as = c("none", "total", "domain"),
  productivity_loss_infection = 1
)

Arguments

x

A <daedalus_output> object from a call to daedalus().

summarise_as

A string from among "none", "total", or "domain", for how the costs should be returned. Select "none", the default, for the raw costs along with overall and domain-specific totals; "total" for the overall cost, and "domain" for the total costs per domain; the domains are 'economic', 'education', and 'life years'.

productivity_loss_infection

A single number in the range \([0, 1]\) giving the loss in productivity associated with symptomatic infection. Currently defaults to 1.0 for compatibility with earlier function versions.

Value

A list of different cost values, including the total cost. See Details for more information.

Details

The total cost in million dollars is returned as total_cost. This is comprised of the following costs.

Economic costs

A three element list of economic_cost_total, the total costs from pandemic impacts on economic sectors, including both costs of lost gross value added (GVA) due to pandemic-control restrictions or closures (economic_cost_closures), and pandemic-related absences due to illness and death (economic_cost_absences).

Educational costs

A three element list of education_cost_total, the total costs from pandemic impacts on education due to pandemic-control restrictions or closures (education_cost_closures), and pandemic-related absences due to illness and death (education_cost_absences).

Life-value lost

A four-element vector (for the number of age groups) giving the value of life-years lost per age group. This is calculated as the life-expectancy of each age group times the value of a statistical life, with all years assumed to have the same value.

Life-years lost

A four-element vector (for the number of age groups) giving the value of life-years lost per age group. This is calculated as the life-expectancy of each age group times the number of deaths in that age group. No quality adjustment is applied.

Examples

output <- daedalus("Canada", "influenza_1918")

get_costs(output)
#> $total_cost
#> [1] 1644627
#> 
#> $economic_costs
#> $economic_costs$economic_cost_total
#> [1] 29545.8
#> 
#> $economic_costs$economic_cost_closures
#> [1] 0
#> 
#> $economic_costs$economic_cost_absences
#> [1] 29545.8
#> 
#> $economic_costs$sector_cost_closures
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [39] 0 0 0 0 0 0 0
#> 
#> $economic_costs$sector_cost_absences
#>  [1]  537.11094   46.96799  951.22060  368.92204  205.81375  534.23991
#>  [7]   38.23981  165.42156  201.57795  172.54020  237.92711   82.37185
#> [13]  161.14994  120.76015  192.11511  232.27379  100.03347   63.77370
#> [19]  246.41226  291.96819  187.33491  262.89409  632.68780  162.51214
#> [25] 2193.78229 3553.29445  800.17533   26.72845  159.12932  351.26550
#> [31]  136.53342  756.61756  300.86158  564.65716  589.36027 2271.80182
#> [37] 3891.57226 1397.03002 1014.27153 2265.94110 1810.27842 2426.75522
#> [43]  257.93605  342.23631   49.58143
#> 
#> 
#> $education_costs
#> $education_costs$education_cost_total
#> [1] 1810.278
#> 
#> $education_costs$education_cost_closures
#> [1] 0
#> 
#> $education_costs$education_cost_absences
#> [1] 1810.278
#> 
#> 
#> $life_value_lost
#> $life_value_lost$life_value_lost_total
#> [1] 1613271
#> 
#> $life_value_lost$life_value_lost_age
#>      0-4     5-19    20-64      65+ 
#> 290122.8 648162.7 444888.8 230096.2 
#> 
#> 
#> $life_years_lost
#> $life_years_lost$life_years_lost_total
#> [1] 35033019
#> 
#> $life_years_lost$life_years_lost_age
#>      0-4     5-19    20-64      65+ 
#>  6300170 14075194  9660994  4996661 
#> 
#>