Vaccine investment scenario parameters

Vaccination parameters for four main pre-pandemic scenarios of pre-pandemic investments in vaccine development and delivery.

vaccination_scenario_names provides the names by which the advance vaccine investment scenarios are known.

Usage

vaccination_scenario_data

vaccination_scenario_names

vaccination_parameter_names

Format

vaccination_scenario_data

A list with 4 elements, each corresponding to an advance vaccine investment strategy. Each element is a list of three values (see Details).

vax_start_time: A single number for the vaccination start time in days.

nu: A single number for the daily vaccination rate as a percentage of the population vaccinated.

vax_uptake_limit: A single number for the percentage of the population that is willing to be vaccinated.

vaccination_scenario_names

A character vector of four elements giving the identifier for each scenario.

vaccination_parameter_names

A character vector of three elements giving the identifier for vaccination parameters.

An object of class character of length 4.

An object of class character of length 3.

Source

See processing details in `data-raw/vaccination_data.R

Details

Each scenario has three parameters; parameters do not differ across age or other groups.

Vaccination start time: The time in days at which vaccination begins.

Vaccination rate: The rate, as a percentage of the population vaccinated per day.

Vaccination uptake limit: The percentage of the population that is willing to accept vaccination.

Examples

# check vaccination scenarios
vaccination_scenario_names
#> [1] "none"   "low"    "medium" "high"  

vaccination_scenario_data
#> $none
#> $none$vax_start_time
#> [1] 365
#> 
#> $none$nu
#> [1] 0.1428571
#> 
#> $none$vax_uptake_limit
#> [1] 40
#> 
#> 
#> $low
#> $low$vax_start_time
#> [1] 300
#> 
#> $low$nu
#> [1] 0.2857143
#> 
#> $low$vax_uptake_limit
#> [1] 50
#> 
#> 
#> $medium
#> $medium$vax_start_time
#> [1] 200
#> 
#> $medium$nu
#> [1] 0.4285714
#> 
#> $medium$vax_uptake_limit
#> [1] 60
#> 
#> 
#> $high
#> $high$vax_start_time
#> [1] 100
#> 
#> $high$nu
#> [1] 0.5
#> 
#> $high$vax_uptake_limit
#> [1] 80
#> 
#>