Validate npi inputs

Validate npi inputs

Usage

validate_npi_input(
  x,
  country,
  infection,
  response_time,
  response_duration = 365
)

Arguments

x

An object to be validated as input to the vaccine_investment argument of daedalus().

country

A country or territory object of class <daedalus_country>, or a country or territory name from those included in the package; see daedalus.data::country_names, or a country ISO2 or ISO3 code; see daedalus.data::country_codes_iso2c and daedalus.data::country_codes_iso3c. Country-specific data such as the community and workplace contacts, the demography, and the distribution of the workforce into economic sectors is automatically accessed from package data for the relevant country name if it is passed as a string. To override package defaults for country characteristics, pass a <daedalus_country> object instead. See daedalus_country() for more.

infection

An infection parameter object of the class <daedalus_infection>, or an epidemic name for which data are provided in the package; see daedalus.data::epidemic_names for historical epidemics or epidemic waves for which parameters are available. Passing the name as a string automatically accesses the default parameters of an infection. Create and pass a <daedalus_infection> to tweak infection parameters using daedalus_infection().

response_time

A single numeric value for the time in days at which the selected response is activated. This is ignored if the response has already been activated by the hospitalisation threshold being reached. Defaults to 30 days. Responses have a default maximum duration of 365 days. This can be changed by passing a <daedalus_npi> object to response_strategy.

Value

A <daedalus_npi> object.