Represent vaccine investment scenarios for DAEDALUS

Helper functions to create and work with S3 class <daedalus_vaccination> objects for use with daedalus(). These objects store vaccination parameters for reuse and have methods for easy parameter access and editing, as well as processing raw vaccination characteristics for the DAEDALUS model.

Usage

daedalus_vaccination(name, ...)

is_daedalus_vaccination(x)

# S3 method for class 'daedalus_vaccination'
print(x, ...)

Arguments

name

A vaccination investment scenario name from among vaccination_scenario_names. Selecting an epidemic automatically pulls in vaccination parameters associated with the epidemic; these are stored as packaged data in daedalus::vaccination_scenario_data. Default vaccination parameters can be over-ridden by passing them as a named list to ....

...

Other parameters passed to print().

x

An object of the <daedalus_vaccination> class.

Vaccination parameters

vax_start_time: The number of days after the start of the epidemic that vaccination begins. Must be a single number.

nu: A single number for the percentage of the total population that can be vaccinated each day. This is converted into a proportion automatically within daedalus().

vax_uptake_limit: A single number giving the upper limit for the percentage of the population that can be vaccinated. When this limit is reached, the vaccination rate nu is set to zero.