DAEDALUS model for health, social and economic costs of a pandemic

Run the DAEDALUS model from R. This is a work in progress.

Usage

daedalus(initial_state, time_end = 300L, parameters = default_parameters())

Arguments

initial_state

The initial state of the population as a numeric array where rows represent age groups, columns represent epidemiological compartments, and layers represent economic sectors.

time_end

An integer-like value for the number of timesteps at which to return data. This is treated as the number of days with data returned for each day. Defaults to 300 days.

parameters

A list of model parameters. See default_parameters() for model defaults.

Value

A <deSolve> object.