DAEDALUS model implemented with dust

daedalus::daedalus2() will eventually replace daedalus::daedalus() with the model implemented in C++ to integrate with dust2 (and in future dust). This is a work in progress!

Usage

daedalus2(
  country,
  infection,
  response_strategy = NULL,
  vaccine_investment = NULL,
  response_time = 30,
  time_end = 100
)

Arguments

country

A country or territory object of class <daedalus_country>, or a country or territory name from those included in the package; see country_names, or a country ISO2 or ISO3 code; see country_codes_iso2c and country_codes_iso3c . Country-specific data such as the community and workplace contacts, the demography, and the distribution of the workforce into economic sectors is automatically accessed from package data for the relevant country name if it is passed as a string. To override package defaults for country characteristics, pass a <daedalus_country> object instead. See daedalus_country() for more.

infection

An infection parameter object of the class <daedalus_infection>, or an epidemic name for which data are provided in the package; see epidemic_names for historical epidemics or epidemic waves for which parameters are available.

Passing the name as a string automatically accesses the default parameters of an infection. Create and pass a <daedalus_infection> to tweak infection parameters using daedalus_infection().

response_strategy

A string for the name of response strategy followed; defaults to "none". The response strategy determines the country-specific response threshold following which the response is activated. See response_threshold.

While the response strategy is active, economic contacts are scaled using the package data object daedalus::closure_data.

vaccine_investment

Either a single string or a <daedalus_vaccination> object specifying the vaccination parameters associated with an advance vaccine-investment scenario. Defaults to "none", for no advance vaccine investment. In this case, vaccination begins 365 days (1 year) after the simulation begins, at a low rate across all age groups. Other accepted values are "low", "medium" and "high". See daedalus_vaccination() for more information.

response_time

A single numeric value for the time in days at which the selected response is activated. This is ignored if the response has already been activated by the hospitalisation threshold being reached. Defaults to 30 days.

time_end

An integer-like value for the number of timesteps at which to return data. This is treated as the number of days with data returned for each day. Defaults to 300 days.

Details

Note that daedalus2() currently uses a default vaccination strategy of no vaccination, using the internal helper function dummy_vaccination(). This is expected to become the default for daedalus() when it is replaced with daedalus2().

Note also that daedalus2() vaccination begins at the model start time, and not at the time specified in vaccine_investment. This functionality will be updated soon.

Examples

# not yet ready for conversion to `<daedalus_output>`
output <- daedalus2("GBR", "sars_cov_1")

names(output)
#> [1] "total_time"           "model_data"           "country_parameters"  
#> [4] "infection_parameters" "response_data"