Get epidemic costs from a DAEDALUS model run

Get epidemic costs from a DAEDALUS model run

Usage

get_costs(x, summarise_as = c("none", "total", "domain"))

Arguments

x

A <daedalus_output> object from a call to daedalus().

summarise_as

A string from among "none", "total", or "domain", for how the costs should be returned. Select "none", the default, for the raw costs along with overall and domain-specific totals; "total" for the overall cost, and "domain" for the total costs per domain; the domains are 'economic', 'education', and 'life years'.

Value

A list of different cost values, including the total cost. See Details for more information.

Details

The total cost in million dollars is returned as total_cost. This is comprised of the following costs.

Economic costs

A three element list of economic_cost_total, the total costs from pandemic impacts on economic sectors, including both costs of lost gross value added (GVA) due to pandemic-control restrictions or closures (economic_cost_closures), and pandemic-related absences due to illness and death (economic_cost_absences).

Educational costs

A three element list of education_cost_total, the total costs from pandemic impacts on education due to pandemic-control restrictions or closures (education_cost_closures), and pandemic-related absences due to illness and death (education_cost_absences).

Life-years lost

A four-element vector (for the number of age groups) giving the value of life-years lost per age group. This is calculated as the life-expectancy of each age group times the value of a statistical life, with all years assumed to have the same value.

Examples

output <- daedalus("Canada", "influenza_1918")

get_costs(output)
#> $total_cost
#> [1] 1451331
#> 
#> $economic_costs
#> $economic_costs$economic_cost_total
#> [1] 41080.77
#> 
#> $economic_costs$economic_cost_closures
#> [1] 0
#> 
#> $economic_costs$economic_cost_absences
#> [1] 41080.77
#> 
#> 
#> $education_costs
#> $education_costs$education_cost_total
#> [1] 2553.988
#> 
#> $education_costs$education_cost_closures
#> [1] 0
#> 
#> $education_costs$education_cost_absences
#> [1] 2553.988
#> 
#> 
#> $life_years_lost
#> $life_years_lost$life_years_lost_total
#> [1] 1407696
#> 
#> $life_years_lost$life_years_lost_age
#>      0-4     5-19    20-65      65+ 
#> 183926.7 554960.9 521622.1 147186.8 
#> 
#>